refactor: extract admin event filters hook
